+++ packages/iov-rise/src/riseconnection.ts
@@ -303,11 +303,6 @@ export class RiseConnection implements BcpConnection {
     throw new Error("Not implemented");
   }
 
-  /** @deprecated use watchBlockHeaders().map(header => header.height) */
-  public changeBlock(): Stream<number> {
-    return this.watchBlockHeaders().map(header => header.height);
-  }
-
   public async searchTx(query: BcpTxQuery): Promise<ReadonlyArray<ConfirmedTransaction>> {
     if (query.height || query.tags) {
       throw new Error("Query by height and tags not supported");